requirements.txt
pytidylib>=0.2.4

nose>=1.3.7
